Beginning of test for billion laughs in webvulnscan
