+++ webvulnscan/attacks/__init__.py
@@ -7,7 +7,8 @@ from .clickjack import clickjack
 from .cookiescan import cookiescan
 from .exotic_characters import exotic_characters
 from .utf7_check import utf7_check
+from .billion_laughs import billion_laughs
 
 
 def all_attacks():
-    return [xss, csrf, crlf, breach, clickjack, cookiescan, exotic_characters,utf7_check]
+    return [xss, csrf, crlf, breach, clickjack, cookiescan, exotic_characters,utf7_check,billion_laughs]
